requirements.txt
attrs==19.3.0
dnspython==1.16.0
importlib-metadata==1.6.0
mock==4.0.2
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
six==1.14.0
wcwidth==0.1.9
zipp==3.1.0
